requirements.txt
# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-play-scraper==1.2.4

# AI/LLM
openai==1.12.0
anthropic==0.18.1
langchain==0.1.9
langchain-openai==0.0.6
langchain-anthropic==0.1.4

# Data Processing
pandas==2.2.0
numpy==1.26.4
python-dateutil==2.8.2

# Utilities
pydantic==2.6.1
tqdm==4.66.1
tenacity==8.2.3

# Storage
jsonlines==4.0.0
